+++ src/components/BatchAllocations/BatchAllocationsForm/utils.js
@@ -1,4 +1,10 @@
-import { BATCH_ALLOCATION_COLUMNS } from '../constants';
+import omit from 'lodash/omit';
+
+import {
+  BATCH_ALLOCATION_COLUMNS,
+  BATCH_ALLOCATION_FIELDS,
+  BATCH_ALLOCATION_FORM_SPECIAL_FIELDS,
+} from '../constants';
 
 export const getBatchAllocationColumnMapping = ({ intl }) => {
   return BATCH_ALLOCATION_COLUMNS.reduce((acc, column) => {
@@ -9,4 +15,11 @@ export const getBatchAllocationColumnMapping = ({ intl }) => {
 export const getFormattedOptions = (intl, dataOptions) => dataOptions.map(({ labelId, ...rest }) => ({
   label: intl.formatMessage({ id: labelId }),
   ...rest,
+}));
+
+export const normalizeFInanceData = (fyFinanceData) => fyFinanceData.map((item) => ({
+  ...omit(item, [BATCH_ALLOCATION_FORM_SPECIAL_FIELDS._isMissed]),
+  [BATCH_ALLOCATION_FIELDS.budgetAllocationChange]: parseFloat(item[BATCH_ALLOCATION_FIELDS.budgetAllocationChange]),
+  [BATCH_ALLOCATION_FIELDS.budgetAllowableExpenditure]: parseFloat(item[BATCH_ALLOCATION_FIELDS.budgetAllowableExpenditure]),
+  [BATCH_ALLOCATION_FIELDS.budgetAllowableEncumbrance]: parseFloat(item[BATCH_ALLOCATION_FIELDS.budgetAllowableEncumbrance]),
 }));