+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Budget batch allocation screen from CSV load. Refs UIF-528.
 * Recalculate function on batch allocation screen. Refs UIF-533.
 * *BREAKING* Batch allocation logs view and delete feature. Refs UIF-530.
+* Fixes for batch allocation form. Refs UIF-579.
 
 ## [7.0.1](https://github.com/folio-org/ui-finance/tree/v7.0.1) (2024-12-02)
 [Full Changelog](https://github.com/folio-org/ui-finance/compare/v7.0.0...v7.0.1)